Chrono Cross - The Radical Dreamers Edition
Redid codes for the game, there was a limitation in how the 0x5 Code Type works, so had to come with a way to get around it.
At the same time, realized that the game wasn't entirely static, so had to again come up with a work around.

Thanks to GamerJin and TomVita for helping me